Scale tree growth by volume of nodes
Also fix leaf decay radii for Celery Top Pine and Australian Cherry.

The growth rate is scaled based on the number of trunks, leaves and
fruits on average for each of the tree's schematics. Leaves and fruit
weigh 1/3 the value of a trunk node; this is called the tree mass.
Tree-mass is then scaled nonlinearly to make smaller trees grow
relatively faster than larger trees; the function was made with the help
of GeoGebra.